requirementsAll.txt
aenum==2.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altgraph==0.17.4
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
crayons==0.4.0
cryptography==43.0.3
decorator==4.4.2
discord.py==2.5.2
fake-useragent==2.2.0
fbchat==1.9.7
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
git-filter-repo==2.47.0
h11==0.16.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
joblib==1.4.2
keyboard==0.13.5
MouseInfo==0.1.3
moviepy==1.0.3
multidict==6.4.3
mutagen==1.47.0
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
paho-mqtt==1.6.1
pefile==2023.2.7
pexpect==4.9.0
piexif==1.1.3
pillow==10.4.0
proglog==0.1.10
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
# PyAudio @ file:///C:/Users/65011/Downloads/PyAudio-0.2.14-cp312-cp312-win_amd64.whl#sha256=12f2f1ba04e06ff95d80700a78967897a489c05e093e3bffa05a84ed9c0a7fa3
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
pydub==0.25.1
PyExifTool==0.5.6
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
# pyqt5-tools @ file:///C:/Users/65011/Downloads/pyqt5_tools-5.15.9.3.3-py3-none-any.whl#sha256=4b45b26111c583a34fa364b98d7120f5aeca4f96ac72e251f1b37a9cca809f86
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.1.0
python-gnupg==0.5.4
python-magic==0.4.27
python-magic-bin==0.4.14
pytweening==1.2.0
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.33.0
selenium-stealth==1.0.6
setuptools==80.9.0
sip==6.11.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
stegano==0.11.4
threadpoolctl==3.5.0
tqdm==4.67.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.0
wsproto==1.2.0
yarl==1.19.0
